reCAPTCHA V2 Added to Login
Added the reCAPTCHA to the login.

created .env file to store the Site Key and Secret Key. Made changes to frontend dockerfile, loginmodal (to add the recaptcha), page.tsx, and I had an error for a while and when I searched it, I was told to suppress a hydration warning. As soon as I did that it started working but the change was added in layout.txt

Keeping these links as a reference for future review (sites I used to set it up):
https://developers.google.com/recaptcha/docs/display